+++ sys/net/routing/ara/evaporation.h
@@ -40,7 +40,7 @@ extern "C" {
  * @param[in] threshold The threshold denotes the smallest allowed pheromone 
  * value before it is set to 0
  */
-void ara_evaporation_init(float evaporation_factor, float threshold, uint16_t time_interval_millis, uint8_t type);
+void ara_evaporation_init(float evaporation_factor, float threshold, uint64_t time_interval_millis, uint8_t type);
 
 /**
  * The method evaporates pheromones in a linear fashion. This means that
@@ -59,9 +59,9 @@ void ara_evaporation_init(float evaporation_factor, float threshold, uint16_t ti
  * @return The new pheromone value after the evaporation function has beend
  * applied.
  */
-float ara_evaporation_linear(float old_pheromone_value, uint16_t milliseconds_since_last_evaporation);
+float ara_evaporation_linear(float old_pheromone_value, uint64_t milliseconds_since_last_evaporation);
 
-float ara_evaporation_exponential(float old_pheromone_value, uint16_t milliseconds_since_last_evaporation);
+float ara_evaporation_exponential(float old_pheromone_value, uint64_t milliseconds_since_last_evaporation);
 
 #ifdef __cplusplus
 }
